Add new `theme_extension` and `language_extension` crates for registration purposes
