+++ crates/extensions_ui/Cargo.toml
@@ -17,13 +17,11 @@ client.workspace = true
 collections.workspace = true
 db.workspace = true
 editor.workspace = true
-extension.workspace = true
 extension_host.workspace = true
 fs.workspace = true
 fuzzy.workspace = true
 gpui.workspace = true
 language.workspace = true
-lsp.workspace = true
 num-format.workspace = true
 picker.workspace = true
 project.workspace = true
