+++ crates/theme_extension/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "theme_extension"
+version = "0.1.0"
+edition = "2021"
+publish = false
+license = "GPL-3.0-or-later"
+
+[lints]
+workspace = true
+
+[lib]
+path = "src/theme_extension.rs"
+
+[dependencies]
+anyhow.workspace = true
+extension.workspace = true
+fs.workspace = true
+gpui.workspace = true
+theme.workspace = true
