+++ crates/zed/Cargo.toml
@@ -59,6 +59,7 @@ inline_completion_button.workspace = true
 install_cli.workspace = true
 journal.workspace = true
 language.workspace = true
+language_extension.workspace = true
 language_model.workspace = true
 language_models.workspace = true
 language_selector.workspace = true
@@ -107,6 +108,7 @@ tasks_ui.workspace = true
 telemetry_events.workspace = true
 terminal_view.workspace = true
 theme.workspace = true
+theme_extension.workspace = true
 theme_selector.workspace = true
 time.workspace = true
 toolchain_selector.workspace = true
